requirements.txt
python-pptx>=0.6.21
pydantic>=2.0
pillow>=10.0
numpy>=1.24
scikit-learn>=1.3
unstructured[pptx]>=0.10
openai>=1.0
azure-identity>=1.15
